net: fix rx queue freeze
Under heavy ingress traffic we might end up in a situation where the tap
device is full, but the rx.deferred_frame flag is not set. Beeing in
this state the firecracker RX queue becomes effectively blocked. The tap
won't issue any new event because of the EPOLLET flag, while any rate
limiter or RX queue event won't do any processing because the
rx.deferred_frame flag is not set.

We need to make sure we set the rx.deferred_frame flag on all the paths
that leave unprocessed frames in the tap.

Signed-off-by: Serban Iorga <seriorga@amazon.com>
Signed-off-by: Ioana Chirca <chioana@amazon.com>