docs: add documentation on snapshot uniqueness
When snapshots are used in a such a manner that a given guest's state is
resumed from more than once, guest information assumed to be unique may
in fact not be; this information can include identifiers, random numbers
and random number seeds, as well as cryptographic tokens. Without a
strong mechanism that enables users to guarantee that unique things stay
unique across snapshot restores, we consider resuming execution from the
same state more than once insecure.

Signed-off-by: Radu Weiss <raduweis@amazon.com>