rpc_interface: tests: add mock VmResources
The RPC interface code has the purpose of transforming locally
defined RPC commands (VmmActions) to programmatic API calls to
the underlying components.

In order to efficiently test the RPC interface code we could create
mock objects for these underlying components and make use of these
mock objects to control and validate expected RPC interface behavior.

This commit adds `struct MockVmRes` which, when under config(test), is
exported as `VmResources`.

With a mock VmResources struct we can easily test most of the
PrebootApiController.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>