vmm: move `mem_size_mib` from persist to vmm
The function that returns the size of the guest memory was used only
in the snapshot functionality of the `vmm` crate, which made it
inaccessible in `vmm/lib.rs` on `aarch64`. Now that the balloon
device needs to poll the size of guest memory, this function is no
longer relevant only to the `persist` module.

Signed-off-by: George Pisaltu <gpl@amazon.com>