tests: updated net latency baseline
This network latency test presents large variability
when run on 4.14 AL2 host kernel. Running it on 5.4
AL2 kernel outputs more stable results, reflected by
the change in the delta, but roughly doubled latency.
This is not concerning taking into account that the test
sole purpose is to detect performance degradations which
can be seen during network latency measurements.

Signed-off-by: Iulian Barbu <iul@amazon.com>